refactor: rename Profiler to TraceHandler to better reflect its responsibility
The component previously named "Profiler" has been renamed to "TraceHandler" to more accurately describe its primary function - handling and processing stack traces rather than managing the entire profiling workflow. This change improves code clarity by:

- Moving the implementation to a dedicated trace_handler.rs file
- Adding blazesym Symbolizer as a component of TraceHandler
- Updating all references throughout the codebase
